requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
django-heroku==0.3.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
gunicorn==23.0.0
numpy==2.1.3
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
whitenoise==6.8.2
